+++ src/schick/seg035.h
@@ -12,7 +12,7 @@ void FIG_loot_monsters(void);
 void FIG_split_ap(void);
 
 /* 0x2f */
-signed short check_hero_range_attack(const struct struct_hero*, const signed int);
+signed int check_hero_range_attack(const struct struct_hero*, const signed int);
 
 #if !defined(__BORLANDC__)
 }
